Decode CodeSource-Path if not found. When running tests in maven the command clazz.getProtectionDomain().getCodeSource().getLocation() contains url-encoded parts. (running from Intellij or eclipse doesnt do)
e.g. if the CodeSource of the requested Class is in
"/opt/abc@xyz/abc.jar" then it will return "/opt/abc%40xyz/abc.jar",
which is a file that cannot be found.
With this fix, if the file does not exist, we try to url-decode the
path.

Before this fix, we had <new File("/opt/abc%40xyz/abc.jar")>.
Then we called .toURI() which again url-encodes the path to
"/opt/abc%2540xyz/abc.jar".

So we have a double-url-encoded path.

Our local maven repo has a @ sign in its path, which leads to a warning,
that the deltaspike-jars could not be loaded (because of this
<urlEncode(urlEncode(path))>) and finally no transactions could be
started, because deltaspike was not loaded.